Move expectPdfToMatchSnapshot.ts to separate folder, as files in \tests should contain only files with tests
